replace prison capsule animals with beer!
the way this works is that, when the switch is hit, we load into vram
the beer sprite in place of where the rabbit sprite would be.
additionally, the animal spawning code checks if controls have been
disabled (ie. level is over), and if so, uses palette line 1 for new
sprites and skips randomization, always spawning "rabbits".